Add Trigger Proxy
Only Trigger results should be saved per event (something like trigger
bits). On the contrary any shared information (such as Trigger Name)
should be kept in a central place that is TriggerProxy.

This way user should obtain TriggerProxy only once and then use this
object in each event to get access to the trigger results.

At the end of the day there is no search for the trigger in each event.