Remove Input type from TreeInfo; Add Version print
TreeInfo input type is not used and leads to the confusion. It has to be
set in each CMSSW job explicitly. This feature is posponed for future
releases. Hopefuly, there will be a way to determine Input type on the
fly.

Version received an output operator that allows to stream version number
to std::ostream