Put Hash typedef adn Generator in Tools
The original idea was to keep a central typedef of the Hash and its
Generator that will be used outside the Tree, Trigger and TriggerCenter.

At the moment CMSSW compilation fails if Tools are included in the
s8::Trigger due to boost 1_42 version. Hopefuly, thigs will change
in near future.